Pokedex Item View
Started working on UI elements by creating a fairly basic PokedexItem which will be displayed as a list. Also created Chip component which can be used for displaying chips (in this case it will be used to show the types). Also started to create a basic model for the different pokemon types.

None of this is wired up to the backend API... yet.